+++ SettingAppMVC/Resources/Assets.xcassets/AccentColor.colorset/Contents.json
