+++ SettingAppMVC/Resources/Assets.xcassets/Contents.json
