+++ SettingAppMVC/Sources/App/AppDelegate.swift
@@ -1,0 +1,18 @@
+//
+//  AppDelegate.swift
+//  SettingAppMVC
+//
+//  Created by Nick Spravtsev on 15.09.2022.
+//
+
+import UIKit
+
+@main
+class AppDelegate: UIResponder, UIApplicationDelegate {
+
+    func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
+        // Override point for customization after application launch.
+        return true
+    }
+}
+
