+++ SettingAppMVC/Sources/App/SceneDelegate.swift
@@ -1,0 +1,21 @@
+//
+//  SceneDelegate.swift
+//  SettingAppMVC
+//
+//  Created by Nick Spravtsev on 15.09.2022.
+//
+
+import UIKit
+
+class SceneDelegate: UIResponder, UIWindowSceneDelegate {
+
+    var window: UIWindow?
+
+    func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
+        guard let windowScene = (scene as? UIWindowScene) else { return }
+        let window = UIWindow(windowScene: windowScene)
+        window.rootViewController = ViewController()
+        window.makeKeyAndVisible()
+        self.window = window
+    }
+}
